re-org loadbalancer tests
